Supervisor (#11)
* Created create-feedback component and feedback service

* Update application.properties

* Implemented backend for Post API

implemented backend methods to save feedback in database

* Implemented component and service class

* post api

* Routing changes

* Implemented fetch employees of associated supervisor

* implemented method to fetch reportees and feedbacks

* Implemented fetch employee method

* implemented saveFeedback at frontend

* implemented frontend methods for fetching all reportees and method for saving feedback

* implemented frontend for view feedback given by me , view my feedbacks

implemented frontend for view feedbacks given for me and view feedback given by me and also implemented search employee by name

* Resolve conflict

* Implemented the ion-segment and search employee by name

* Implemented methods for accessing api using httpclient

* Implemented back button and refactored code

* Implemented view feedback functionality

* Implemented Controller, Service and Repository methods

Implemented Controller, Service and Repository methods for View my feedback, view feedback given by me, delete feedback and update feedback.
Also implemented one one mapping.

* Added validation to form, alerts and toast.

* Implemented one to one mapping to fetch name of reportees

Implemented methods and one to one mapping to fetch feedback and user data

* implemented rest api for update feedback and fetch feedback by using id

* implemented alert for delete

* Implemented and resolved errors encountered while saving the feedback

* Implemented frontend to fetch reportee and supervisor name

* Fetched reportee name and supervisor name from employee model through feedback model

* created ddl schemas for database

* Tost service to display acknowledgement

* component class for add feedback

* made required changes , changed the rest api paths and use changed naming conventions at backend

* made required changes

* Implemented pagination on client side and server side

* required changes

* Given validations for success and error for add, delete and update the feedbacks

* Used Query Parameter for URL

* implements role based UI component view

* made changes for accessibility of users

* added error handlers

* Checked user is supervisor at frontend

* made changes

* Exception and authorization checked

* required changes

* Provided comments and formatting

* Required changes

* Delete toast.service.ts

* done required changes

* Update feedback-list.component.html

* Update .gitignore

* Revert "Update application.properties"

This reverts commit ffec9afec28b6335f391cc108c776ab874dfbcf7.

* Added error messages to error handlers

* required changes

* Revert "Tost service to display acknowledgement"

This reverts commit 9f72ad3879f05be62d554c49131633f280ab401d.

* made PR changes at backend

* changed urls

* changed the naming conventions in entity

* Required changes

* component deleted

* added icons

* null check done

---------

+++ Server/src/main/java/com/abs/errp/security/LoggedInUser.java
@@ -8,10 +8,10 @@ import org.springframework.security.core.userdetails.User;
 public class LoggedInUser extends User {
 	static final long serialVersionUID = 1L;
 
-	private Long employeeId;
+	private int employeeId;
 	private String employeeName;
 	
-	public LoggedInUser(String username, String password, Long employeeId, String employeeName, Collection<? extends GrantedAuthority> authorities) {
+	public LoggedInUser(String username, String password, int employeeId, String employeeName, Collection<? extends GrantedAuthority> authorities) {
 		super(username, password, authorities);
 		this.employeeName=employeeName;
 		this.employeeId = employeeId;
@@ -25,11 +25,11 @@ public class LoggedInUser extends User {
 		this.employeeName = employeeName;
 	}
 
-	public Long getEmployeeId() {
+	public int getEmployeeId() {
 		return employeeId;
 	}
 
-	public void setEmployeeId(Long employeeId) {
+	public void setEmployeeId(int employeeId) {
 		this.employeeId = employeeId;
 	}
 }
